requirements.txt
docopt
onkyo-eiscp
pyserial
pyxdg
Twisted
wheel
zope.interface
sphinx
Babel
snowballstemmer
sphinx-rtd-theme
six
Pygments
Jinja2
alabaster
pytz
MarkupSafe
docutils
mock
